requirements.txt
asgiref==3.2.10
Django==3.0.7
django-admin-rangefilter==0.6.1
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
